Initial move and refactor MediaControls
