+++ packages/apollos-ui-connected/src/MediaControlsConnected/WebView.js
@@ -1,0 +1,24 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+
+import { RockAuthedWebBrowser } from '@apollosproject/ui-connected';
+
+import PlayButton from './PlayButton';
+
+const WebView = ({ webViewUrl, coverImageSources }) => (
+  <RockAuthedWebBrowser>
+    {(openUrl) => (
+      <PlayButton
+        action={() => openUrl(webViewUrl)}
+        coverImageSources={coverImageSources}
+      />
+    )}
+  </RockAuthedWebBrowser>
+);
+
+WebView.propTypes = {
+  webViewUrl: PropTypes.func,
+  coverImageSources: PropTypes.arrayOf(PropTypes.string),
+};
+
+export default WebView;
