+++ packages/apollos-ui-connected/src/MediaControlsConnected/index.js
@@ -1,0 +1,3 @@
+import MediaControlsConnected from './MediaControlsConnected';
+
+export default MediaControlsConnected;
